refactor: rename and relocate TranslationSanitizationParameters to DTO ss-044
